splits email utils
