+++ services/web/server/src/simcore_service_webserver/login/utils.py
@@ -1,33 +1,20 @@
 import logging
-import mimetypes
 import random
-from email import encoders
-from email.mime.base import MIMEBase
-from email.mime.multipart import MIMEMultipart
-from email.mime.text import MIMEText
-from os.path import join
-from pathlib import Path
-from pprint import pformat
-from typing import Any, Mapping, Optional
+from typing import Any, Optional
 
-import aiosmtplib
 import attr
 import passlib.hash
 from aiohttp import web
-from aiohttp_jinja2 import render_string
 from models_library.users import UserID
 from passlib import pwd
 from servicelib import observer
 from servicelib.aiohttp.rest_models import LogMessageType
 from servicelib.json_serialization import json_dumps
 from servicelib.mimetype_constants import MIMETYPE_APPLICATION_JSON
-from settings_library.email import EmailProtocol
 from simcore_postgres_database.models.users import UserRole
-from simcore_service_webserver.products import get_product_template_path
 
-from .._resources import resources
 from ..db_models import ConfirmationAction, UserRole, UserStatus
-from .settings import LoginOptions, get_plugin_options
+from .settings import LoginOptions
 
 log = logging.getLogger(__name__)
 
@@ -54,7 +41,7 @@ REGISTRATION, RESET_PASSWORD, CHANGE_EMAIL = _to_names(
 )
 
 
-def validate_user_status(user: dict, cfg, support_email: str):
+def validate_user_status(user: dict, cfg: LoginOptions, support_email: str):
     user_status: str = user["status"]
 
     if user_status == BANNED or user["role"] == ANONYMOUS:
@@ -115,70 +102,6 @@ def get_client_ip(request: web.Request) -> str:
     return ips.split(",")[0]
 
 
-async def compose_mail(
-    app: web.Application, recipient: str, subject: str, body: str
-) -> None:
-    msg = MIMEText(body, "html")
-    msg["Subject"] = subject
-    msg["To"] = recipient
-
-    await send_mail(app, msg)
-
-
-async def compose_multipart_mail(
-    app: web.Application,
-    recipient: str,
-    subject: str,
-    body: str,
-    attachments: list[tuple[str, bytearray]],
-) -> None:
-    msg = MIMEMultipart("alternative")
-    msg["Subject"] = subject
-    msg["To"] = recipient
-
-    part1 = MIMEText(body, "html")
-    msg.attach(part1)
-
-    for attachment in attachments:
-        filename = attachment[0]
-        payload = attachment[1]
-        mimetype = mimetypes.guess_type(filename)[0].split("/")
-        part = MIMEBase(mimetype[0], mimetype[1])
-        part.set_payload(payload)
-        part.add_header("Content-Disposition", f'attachment; filename="{filename}"')
-        encoders.encode_base64(part)
-        msg.attach(part)
-
-    await send_mail(app, msg)
-
-
-async def render_and_send_mail(
-    request: web.Request,
-    to: str,
-    template: Path,
-    context: Mapping[str, Any],
-    attachments: Optional[list[tuple[str, bytearray]]] = None,
-):
-    page = render_string(template_name=f"{template}", request=request, context=context)
-    # NOTE: Expects first line of the template to be the Subject of the email
-    subject, body = page.split("\n", 1)
-
-    if attachments:
-        await compose_multipart_mail(
-            request.app, to, subject.strip(), body, attachments
-        )
-    else:
-        await compose_mail(request.app, to, subject.strip(), body)
-
-
-def themed(dirname, template) -> Path:
-    return resources.get_path(join(dirname, template))
-
-
-async def get_template_path(request: web.Request, filename: str) -> Path:
-    return await get_product_template_path(request, filename)
-
-
 def flash_response(
     message: str, level: str = "INFO", *, status: int = web.HTTPOk.status_code
 ) -> web.Response:
@@ -201,45 +124,3 @@ def envelope_response(
         status=status,
     )
     return rsp
-
-
-async def send_mail(app: web.Application, msg: MIMEText):
-    cfg: LoginOptions = get_plugin_options(app)
-    log.debug("Email configuration %s", cfg)
-
-    msg["From"] = cfg.SMTP_SENDER
-    smtp_args = dict(
-        hostname=cfg.SMTP_HOST,
-        port=cfg.SMTP_PORT,
-        use_tls=cfg.SMTP_PROTOCOL == EmailProtocol.TLS,
-        start_tls=cfg.SMTP_PROTOCOL == EmailProtocol.STARTTLS,
-    )
-    log.debug("Sending email with smtp configuration: %s", pformat(smtp_args))
-    if cfg.SMTP_PORT == 587:
-        # NOTE: aiosmtplib does not handle port 587 correctly
-        # this is a workaround
-        smtp = aiosmtplib.SMTP(**smtp_args)
-        if cfg.SMTP_PROTOCOL == EmailProtocol.STARTTLS:
-            log.info("Unencrypted connection attempt to mailserver ...")
-            await smtp.connect(use_tls=False, port=cfg.SMTP_PORT)
-            log.info("Starting STARTTLS ...")
-            await smtp.starttls()
-        elif cfg.SMTP_PROTOCOL == EmailProtocol.TLS:
-            await smtp.connect(use_tls=True, port=cfg.SMTP_PORT)
-        elif cfg.SMTP_PROTOCOL == EmailProtocol.UNENCRYPTED:
-            await smtp.connect(use_tls=False, port=cfg.SMTP_PORT)
-            log.info("Unencrypted connection attempt to mailserver ...")
-        #
-        if cfg.SMTP_USERNAME and cfg.SMTP_PASSWORD:
-            log.info("Attempting a login into the email server ...")
-            await smtp.login(cfg.SMTP_USERNAME, cfg.SMTP_PASSWORD.get_secret_value())
-        await smtp.send_message(msg)
-        await smtp.quit()
-    else:
-        async with aiosmtplib.SMTP(**smtp_args) as smtp:
-            if cfg.SMTP_USERNAME and cfg.SMTP_PASSWORD:
-                log.info("Login email server ...")
-                await smtp.login(
-                    cfg.SMTP_USERNAME, cfg.SMTP_PASSWORD.get_secret_value()
-                )
-            await smtp.send_message(msg)
