+++ services/web/server/tests/unit/with_dbs/03/test_login_utils.py
@@ -16,7 +16,7 @@ from json2html import json2html
 from pytest_mock import MockerFixture
 from simcore_service_webserver._constants import RQ_PRODUCT_KEY
 from simcore_service_webserver.email import setup_email
-from simcore_service_webserver.login.utils import (
+from simcore_service_webserver.login.utils_email import (
     get_template_path,
     render_and_send_mail,
     themed,
